perf(hyperdim): optimize AVX2 hamming distance and enhance verification
- Hoisted zero-vector intrinsic out of the hot loop in `hamming_distance_simd_avx2`.
- Unrolled loop 2x to process 512 bits per iteration, reducing loop control overhead.
- Implemented dual independent accumulators to break dependency chains and improve ILP.
- Added technical documentation verifying 80-word alignment for the unroll factor.
- Expanded correctness tests with 100 random iterations comparing SIMD vs naive bit-by-bit reference.
- Verified ~19.1% latency improvement in `cosine_similarity` benchmark.
