--- .jules/paper-research-last-run.txt
@@ -1,1 +1,0 @@
-2026-04-20
