+++ llms.txt
@@ -54,7 +54,7 @@ Created by: [cargo-llms-txt](https://github.com/masinc/cargo-llms-txt)
 ### src/bin/csm.rs
 
 - pub use native::chaotic_semantic_memory::cli::commands::watch::EventFilter
-- pub use native::chaotic_semantic_memory::cli::{CliArgs, CliError, Commands, CompletionsArgs, ExitCode, OutputFormat, ensure_git_local_dir, resolve_git_local_path, run_associate, run_associations, run_completions, run_delete, run_diff, run_disassociate, run_export, run_get, run_history, run_import, run_index_dir, run_index_jsonl, run_inject, run_metrics, run_path, run_probe, run_probe_filtered, run_probe_graph, run_query, run_rollback, run_stats, run_traverse, run_update, run_watch}
+- pub use native::chaotic_semantic_memory::cli::{CliArgs, CliError, Commands, CompletionsArgs, ExitCode, OutputFormat, ensure_git_local_dir, resolve_git_local_path, run_associate, run_associations, run_completions, run_delete, run_disassociate, run_export, run_get, run_import, run_index_dir, run_index_jsonl, run_inject, run_metrics, run_path, run_probe, run_probe_filtered, run_probe_graph, run_query, run_stats, run_traverse, run_update, run_watch}
 - pub use native::clap::Parser
 - pub use native::colored::Colorize
 - pub use native::tracing::Level
@@ -110,9 +110,6 @@ Created by: [cargo-llms-txt](https://github.com/masinc/cargo-llms-txt)
 - pub struct MetricsArgs
 - pub struct WatchArgs
 - pub struct ProbeGraphArgs
-- pub struct HistoryArgs
-- pub struct DiffArgs
-- pub struct RollbackArgs
 
 ### src/cli/commands/associate.rs
 
@@ -131,10 +128,6 @@ Created by: [cargo-llms-txt](https://github.com/masinc/cargo-llms-txt)
 
 - pub fn run_delete
 
-### src/cli/commands/diff.rs
-
-- pub fn run_diff
-
 ### src/cli/commands/disassociate.rs
 
 - pub fn run_disassociate
@@ -146,10 +139,6 @@ Created by: [cargo-llms-txt](https://github.com/masinc/cargo-llms-txt)
 ### src/cli/commands/get.rs
 
 - pub fn run_get
-
-### src/cli/commands/history.rs
-
-- pub fn run_history
 
 ### src/cli/commands/import.rs
 
@@ -179,11 +168,9 @@ Created by: [cargo-llms-txt](https://github.com/masinc/cargo-llms-txt)
 - pub mod associations
 - pub mod completions
 - pub mod delete
-- pub mod diff
 - pub mod disassociate
 - pub mod export
 - pub mod get
-- pub mod history
 - pub mod import
 - pub mod index_dir
 - pub mod index_jsonl
@@ -194,7 +181,6 @@ Created by: [cargo-llms-txt](https://github.com/masinc/cargo-llms-txt)
 - pub mod probe_filtered
 - pub mod probe_graph
 - pub mod query
-- pub mod rollback
 - pub mod stats
 - pub mod traverse
 - pub mod update
@@ -203,11 +189,9 @@ Created by: [cargo-llms-txt](https://github.com/masinc/cargo-llms-txt)
 - pub use associations::run_associations
 - pub use completions::run_completions
 - pub use delete::run_delete
-- pub use diff::run_diff
 - pub use disassociate::run_disassociate
 - pub use export::run_export
 - pub use get::run_get
-- pub use history::run_history
 - pub use import::run_import
 - pub use index_dir::run_index_dir
 - pub use index_jsonl::run_index_jsonl
@@ -218,7 +202,6 @@ Created by: [cargo-llms-txt](https://github.com/masinc/cargo-llms-txt)
 - pub use probe_filtered::run_probe_filtered
 - pub use probe_graph::run_probe_graph
 - pub use query::run_query
-- pub use rollback::run_rollback
 - pub use stats::run_stats
 - pub use traverse::run_traverse
 - pub use update::run_update
@@ -252,10 +235,6 @@ Created by: [cargo-llms-txt](https://github.com/masinc/cargo-llms-txt)
 ### src/cli/commands/query.rs
 
 - pub fn run_query
-
-### src/cli/commands/rollback.rs
-
-- pub fn run_rollback
 
 ### src/cli/commands/stats.rs
 
@@ -304,7 +283,7 @@ Created by: [cargo-llms-txt](https://github.com/masinc/cargo-llms-txt)
 - pub mod git_local
 - pub mod mcp
 - pub use args::*
-- pub use commands::{run_associate, run_associations, run_completions, run_delete, run_diff, run_disassociate, run_export, run_get, run_history, run_import, run_index_dir, run_index_jsonl, run_inject, run_metrics, run_path, run_probe, run_probe_filtered, run_probe_graph, run_query, run_rollback, run_stats, run_traverse, run_update, run_watch}
+- pub use commands::{run_associate, run_associations, run_completions, run_delete, run_disassociate, run_export, run_get, run_import, run_index_dir, run_index_jsonl, run_inject, run_metrics, run_path, run_probe, run_probe_filtered, run_probe_graph, run_query, run_stats, run_traverse, run_update, run_watch}
 - pub use error::{CliError, ExitCode, Result}
 - pub use git_local::{ensure_git_local_dir, resolve_git_local_path}
 - pub use mcp::*
@@ -553,7 +532,7 @@ Created by: [cargo-llms-txt](https://github.com/masinc/cargo-llms-txt)
 - pub use prelude::crate::framework_events::MemoryEvent
 - pub use prelude::crate::hyperdim::HVec10240
 - pub use prelude::crate::semantic_bridge::{BridgeHit, ConceptGraph, MemoryPacket}
-- pub use prelude::crate::singularity::{Concept, ConceptBuilder, ConceptDiff, ConceptVersion}
+- pub use prelude::crate::singularity::{Concept, ConceptBuilder}
 - pub use prelude::crate::singularity_retrieval::{CandidateSource, FilterStrategy, RetrievalConfig, RetrievalStats}
 - pub mod wasm
 
@@ -718,9 +697,6 @@ Created by: [cargo-llms-txt](https://github.com/masinc/cargo-llms-txt)
 - pub struct SingularityConfig
 - impl Default for SingularityConfig
 - pub struct Concept
-- pub struct ConceptVersion
-- pub struct ConceptDiff
-- impl ConceptDiff
 - pub struct ConceptBuilder
 - impl ConceptBuilder
 - pub struct Singularity
