--- progress/paper-research-2026-04-20.md
@@ -1,1 +1,0 @@
-No papers newer than 2026-01-01 were found on any topic.
