+++ src/cli/commands/mod.rs
@@ -2,11 +2,9 @@ pub mod associate;
 pub mod associations;
 pub mod completions;
 pub mod delete;
-pub mod diff;
 pub mod disassociate;
 pub mod export;
 pub mod get;
-pub mod history;
 pub mod import;
 pub mod index_dir;
 pub mod index_jsonl;
@@ -17,7 +15,6 @@ pub mod probe;
 pub mod probe_filtered;
 pub mod probe_graph;
 pub mod query;
-pub mod rollback;
 pub mod stats;
 pub mod traverse;
 pub mod update;
@@ -27,11 +24,9 @@ pub use associate::run_associate;
 pub use associations::run_associations;
 pub use completions::run_completions;
 pub use delete::run_delete;
-pub use diff::run_diff;
 pub use disassociate::run_disassociate;
 pub use export::run_export;
 pub use get::run_get;
-pub use history::run_history;
 pub use import::run_import;
 pub use index_dir::run_index_dir;
 pub use index_jsonl::run_index_jsonl;
@@ -42,7 +37,6 @@ pub use probe::run_probe;
 pub use probe_filtered::run_probe_filtered;
 pub use probe_graph::run_probe_graph;
 pub use query::run_query;
-pub use rollback::run_rollback;
 pub use stats::run_stats;
 pub use traverse::run_traverse;
 pub use update::run_update;
